Move functions to helpers.go
